🚀 Nova: Add missing E2E tests for viral widgets
Adds e2e tests for Viral Receipt Lottery, and updates names/auth context for Job Board and Mystery Offer Generators to match UI.
